requirements.txt
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
dnspython==1.16.0
more-itertools==8.3.0
packaging==20.4
pluggy==0.13.1
py==1.8.1
pycryptodome==3.9.7
pyparsing==2.4.7
pytest==5.4.3
six==1.15.0
wcwidth==0.2.4
